test: split models and use static methods
