+++ cypress/e2e/api-contructor.spec.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
+import ToggleModel from "../support/togglemodel";
 import PageModel from "../support/pagemodel";
-const pageModel = new PageModel();
 
 describe("Toggle set by API", () => {
   context("Given ECMAS bootstrap toggle interface",()=>{
@@ -15,11 +15,11 @@ function testCase(bstInterface) {
   context("When the API constructor method is call", () => {
     const data_test = "api-constructor";
     it("Then toggle is render with the options provided to the API constructor method", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
         cy.wrap($test).find("button").click()
         cy.wrap($test).find('.toggle').then(($toggle)=>{
-          pageModel.checkToggleElementWithOptions($toggle, JSON.parse($test.find('code').html()))
+          ToggleModel.checkToggleElementWithOptions($toggle, JSON.parse($test.find('code').html()))
         });
       });
     });
