+++ cypress/e2e/layout.spec.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
-import PageModel from "../support/pagemodel";
-const pageModel = new PageModel();
+import ToggleModel from "../support/togglemodel";
+import PageModel from "../support/PageModel";
 
 describe("Layout feature", () => {
   context("Given ECMAS bootstrap toggle interface",()=>{
@@ -15,15 +15,15 @@ function testCase(bstInterface) {
   context("When a toggle is in Input Group", () => {
     const data_test = 'layout';
     it("Then toggle height must be equal to elements height with this size", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
-        pageModel.assertComputedStyleProperty(
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
+        PageModel.assertComputedStyleProperty(
           $test,
           "input",
           ".toggle",
           "height"
         );
-        pageModel.assertComputedStyleProperty(
+        PageModel.assertComputedStyleProperty(
           $test,
           "input",
           ".toggle",
@@ -32,9 +32,9 @@ function testCase(bstInterface) {
       });
     });
     it("Then toggle change on click", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
-        pageModel.checkToggleClick($test, true);
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
+        ToggleModel.checkToggleClick($test, true);
       });
     });
   });
