+++ cypress/e2e/sizes.spec.cy.js
@@ -1,6 +1,5 @@
 /// <reference types="cypress" />
 import PageModel from "../support/pagemodel";
-const pageModel = new PageModel();
 
 describe("Size feature", () => {
   context("Given ECMAS bootstrap toggle interface",()=>{
@@ -15,9 +14,9 @@ function testCase(bstInterface) {
   context("When data-size attribute is set", () => {
     const data_test = 'size';
     it("Then toggle height must be equal to button height with this size", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
-        pageModel.assertComputedStyleProperty(
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
+        PageModel.assertComputedStyleProperty(
           $test,
           "button",
           ".toggle",
@@ -27,9 +26,9 @@ function testCase(bstInterface) {
     });
 
     it("Then toggle height must be equal to select height with this size", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
-        pageModel.assertComputedStyleProperty(
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
+        PageModel.assertComputedStyleProperty(
           $test,
           "select",
           ".toggle",
@@ -39,9 +38,9 @@ function testCase(bstInterface) {
     });
 
     it("Then toggle height must be equal to input height with this size", () => {
-      pageModel.load(bstInterface, data_test);
-      pageModel.getTests().each(($test) => {
-        pageModel.assertComputedStyleProperty(
+      PageModel.load(bstInterface, data_test);
+      PageModel.getTests().each(($test) => {
+        PageModel.assertComputedStyleProperty(
           $test,
           'input[type="text"]',
           ".toggle",
