+++ cypress/e2e/status.spec.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
+import ToggleModel from "../support/togglemodel";
 import PageModel from "../support/pagemodel";
-const pageModel = new PageModel();
 
 describe("Status feature", () => {
   context("Given ECMAS bootstrap toggle interface",()=>{
@@ -16,30 +16,30 @@ function testCase(bstInterface) {
   context("When toggle is enabled", () => {
     const test_id = "#status-enabled";
     it("Then toggle opacity is equal to bootstrap enabled button opacity", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','opacity');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','opacity');
       });
     });
 
     it("Then toggle cursor is equal to bootstrap enabled button cursor", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','cursor');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','cursor');
       });
     });
 
     it("Then toggle pointer events are equal to bootstrap enabled button pointer events", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','pointer-events');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','pointer-events');
       });
     });
 
     it("Then toggle change on click", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.checkToggleClick($test, true);
+        ToggleModel.checkToggleClick($test, true);
       });
     });
   });
@@ -47,30 +47,30 @@ function testCase(bstInterface) {
   context("When toggle is disabled", () => {
     const test_id = "#status-disabled";
     it("Then toggle opacity is equal to bootstrap disabled button opacity", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','opacity');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','opacity');
       });
     });
 
     it("Then toggle cursor is equal to bootstrap disabled button cursor", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','cursor');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','cursor');
       });
     });
 
     it("Then toggle pointer events are equal to bootstrap disabled button pointer events", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.assertComputedStyleProperty($test,'button','.toggle','pointer-events');
+        PageModel.assertComputedStyleProperty($test,'button','.toggle','pointer-events');
       });
     });
 
     it("Then toggle do not change on click", () => {
-      pageModel.load(bstInterface, data_test);
+      PageModel.load(bstInterface, data_test);
       cy.get(test_id).each(($test) => {
-        pageModel.checkToggleClick($test, false);
+        ToggleModel.checkToggleClick($test, false);
       });
     });
   });
